Merged HTTP authentication related encoding fix from palday.
Minor stylistic changes made to the original patch.

Reverted version information related changes from the original commit as I do
not have a good enough use case for any problems the current versioning scheme
might be causing and still do not have a clear picture of how we should be
versioning this project's releases and still have them 'fully compatible but not
conflicting' with the original suds project.