Restore last_sent and last_received functions
This revert 8eaf105 with add back the two _SoapClient functions.

Also add back the messages attribute of the client class.

The messages attribute and two functions could be used to get/set
the send/reply message Document of soap request, which could be
used to get the element like 'sessionId' of a soap session.

Updated test_operation_request_and_reply with the messages
attribute and function check.

This resolves issue:
https://github.com/suds-community/suds/issues/79

Signed-off-by: Wayne Sun <gsun@redhat.com>